backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
pymongo
google-generativeai
pydantic
